private-tube-factory: tube ID is a guint64
Fix a crash on 64 bits archs.

https://bugs.freedesktop.org/show_bug.cgi?id=70038

+++ src/private-tubes-factory.c
@@ -996,7 +996,7 @@ channel_closed_cb (GabbleTubeIface *tube,
 {
   GabblePrivateTubesFactoryPrivate *priv =
     GABBLE_PRIVATE_TUBES_FACTORY_GET_PRIVATE (self);
-  guint id;
+  guint64 id;
 
   g_object_get (tube,
       "id", &id,
